in_13

+++ ex00/megaphone.cpp
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   megaphone.cpp                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: brmohamm <brmohamm@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/14 11:34:09 by brmohamm          #+#    #+#             */
+/*   Updated: 2022/09/14 11:34:09 by brmohamm         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <iostream>
 #include <cstring>
 
@@ -7,7 +19,7 @@ int main(int argc, char** argv)
     {
         for(int i = 1 ; i < argc ; i++)
         {
-            for (int x = 0; x < strlen(argv[i]); x++)
+            for (int x = 0; x < (int)strlen(argv[i]); x++)
                     putchar(toupper(argv[i][x]));
         }
     }
